accept a asP31Value parameter on move to wikidata endpoint
in order to push an entity with a different P31
skip updating the entity since it is destroyed after the wd entity
creation
for the moment validate value indistinctivly of types (author can have work P31 values) maybe its enough for the moment as only works and author can be pushed to wikidata and author P31 value is only Q5. But a stricter type validation could be nice to thorw on pushing work P31 as Q5.