prevent sending full errors on endpoints subject to server-side request forgery
By not returning the full errors, we limit the value of such requests to local resources.
An attacker would still be able to trigger a GET request to private network services,
but would not know when such a service was actually found, as the error response
would look the same

This could be enough if we can be sure that nothing wrong can be done by letting
those forged GET requests go through. Additional mitigation could be done
- by running those request in an isolated environment (best security, but increased complexity)
- by additionnaly checking that the targetted ip is a private network ip