verify signature within controller
to allow some activity types to be ignored, without having to check a signature.

This aims to reduce the noise produced by signature verification errors
on remote user Delete requests, which would have no effect anyway.
Those requests trigger a get on the actor endpoint, which answers with a '410 Gone'
as the user was deleted