summaries: include wikipedia sitelinks data
Letting the work of actually fetching the corresponding article extracts to the client for now.
Another implementation could be that there is one endpoint to inform of the available summary providers,
and another endpoint to actually get the desired summary.
The idea to return all summaries does not work well with sitelinks, as some authors can have over a hundred
Wikipedia articles