rename list -> listing
at the exception of the external API which keep the term 'list', to be
coherent the the UI which will also keep list.

Reasonning:
when coding, list is a too general term that is too commonly used in the code
(variables and comments within models, controllers, middlewares, tests
etc.)
listing has the advantange of being close graphically to list, without
having to find synonyms when wanting to talk about a list outside the
users lists described in this commit. As listing improve grepability
(searching for terms in the codebase), we can keep safely the habit of using
list everywhere in the code