remote images: use inventaire.io cached images in development
Without this, opening an inventory browser dropdown selector can trigger some 100 resized images at once to Wikimedia Commons,
which then takes a while to respond, blocking other requests.

An alternative fix would be to implement a cache mechanism for remote images for development (as prod uses Nginx cache)
but that seems to be adding much more complexity that this simple redirection.