get by-suspect-uris endpoint: use a byUriAndType dedicated couchview
makes a reponsability separation between hooks and endpoint. It keeps couch view keys simple (not having byUrisAndbyStateAndByType) while restricts the endpoint access to stateless (aka not archived) tasks. This is possible since the client does not need to access archived tasks for the moment, so the new couch view only returns non-archived tasks.