items: remove snapshot cache
The snapshot refresh triggered by the entity:change event was triggering too many requests to irrelevant entities.
A fix could have been to try to filter relevant entities by checking items, but as the snapshot cache
as been a source of troubles for a while (mostly around invalidation issues), it might be a good opportunity to reconsider the need for such a cache.

In absence of the items snapshot cache, the performance of hot paths such as /api/items?action=by-ids relies on the entities cache.